chore(rolldown_core): rename `BuildState` to `ScanStage`
