+++ crates/rolldown/src/bundler/stages/mod.rs
@@ -1,1 +1,1 @@
-pub mod build_stage;
+pub mod scan_stage;
